requirements.txt
Pillow==6.0.0
django==2.1
django-ckeditor==5.7.1